+++ data/course/src/main/java/com/teamsolply/solply/course/repository/CourseRepositoryImpl.kt
@@ -23,7 +23,7 @@ class CourseRepositoryImpl @Inject constructor(
         )
     }
 
-    override suspend fun getRecommendedCourse(townId: Int) = runCatching {
+    override suspend fun getRecommendedCourse(townId: Long) = runCatching {
         courseRemoteDataSource.getRecommendedCourse(townId).courseList
     }.mapCatching { courseList ->
         courseList.map { course ->
